made it 64bit

+++ device_test/AppleMikeyPaneController.h
@@ -33,7 +33,7 @@
     IBOutlet NSArrayController * mMikeysEventsController;
 
     NSArray * mMikeys;
-    unsigned mMikeyIndex;
+    NSUInteger mMikeyIndex;
     NSMutableArray * mEvents;
     
     // Don't retain these
@@ -43,8 +43,8 @@
 - (NSArray *) mikeys;
 - (void) setMikeys: (NSArray *) theMikeys;
 
-- (unsigned) mikeyIndex;
-- (void) setMikeyIndex: (unsigned) theMikeyIndex;
+- (NSUInteger) mikeyIndex;
+- (void) setMikeyIndex: (NSUInteger) theMikeyIndex;
 
 - (NSMutableArray *) events;
 - (void) setEvents: (NSMutableArray *) theEvents;
